sources/builtin/screen/pyproject.toml
[project]
name = "observer-source-screen"
version = "0.1.0"
description = "Screen capture + OCR source plugin for Observer"
requires-python = ">=3.12"
dependencies = [
    "source-framework",
    "pyobjc-core>=10.0; sys_platform == 'darwin'",
    "pyobjc-framework-Cocoa>=10.0; sys_platform == 'darwin'",
    "pyobjc-framework-Quartz>=10.0; sys_platform == 'darwin'",
    "pyobjc-framework-Vision>=10.0; sys_platform == 'darwin'",
    "Pillow>=10.0",
    "mss>=9.0; sys_platform == 'win32'",
    "rapidocr-onnxruntime>=1.3; sys_platform == 'win32'",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/screen_source"]

[tool.uv.sources]
source-framework = { path = "../../framework" }
